🌱 Extract card metadata and icons from CardWrapper
Move CARD_TITLES (209 entries), CARD_DESCRIPTIONS (155 entries),
and CARD_ICONS (160 entries) from CardWrapper.tsx into dedicated
data files: cardMetadata.ts and cardIcons.ts.

CardWrapper re-exports CARD_TITLES and CARD_DESCRIPTIONS for
backwards compatibility — existing consumers (useSearchIndex,
GPUReservations) continue to import from CardWrapper unchanged.

CardWrapper.tsx: 1,647 → 1,076 lines (-571 lines / -35%)
