+++ web/src/components/cards/cardMetadata.ts
@@ -1,0 +1,370 @@
+// Card title and description registries
+// Pure data — no runtime dependencies
+
+export const CARD_TITLES: Record<string, string> = {
+  // Core cluster cards
+  cluster_health: 'Cluster Health',
+  cluster_focus: 'Cluster Focus',
+  cluster_network: 'Cluster Network',
+  cluster_comparison: 'Cluster Comparison',
+  cluster_costs: 'Cluster Costs',
+  cluster_metrics: 'Cluster Metrics',
+  cluster_locations: 'Cluster Locations',
+  cluster_resource_tree: 'Cluster Resource Tree',
+
+  // Workload and deployment cards
+  app_status: 'Workload Status',
+  workload_deployment: 'Workloads',
+  deployment_missions: 'Deployment Missions',
+  deployment_progress: 'Deployment Progress',
+  deployment_status: 'Deployment Status',
+  deployment_issues: 'Deployment Issues',
+  statefulset_status: 'StatefulSet Status',
+  daemonset_status: 'DaemonSet Status',
+  replicaset_status: 'ReplicaSet Status',
+  job_status: 'Job Status',
+  cronjob_status: 'CronJob Status',
+  hpa_status: 'HPA Status',
+  cluster_groups: 'Cluster Groups',
+  resource_marshall: 'Resource Marshall',
+  workload_monitor: 'Workload Monitor',
+  llmd_stack_monitor: 'llm-d Stack Monitor',
+  prow_ci_monitor: 'PROW CI Monitor',
+  github_ci_monitor: 'GitHub CI Monitor',
+  cluster_health_monitor: 'Cluster Health Monitor',
+
+  // Pod and resource cards
+  pod_issues: 'Pod Issues',
+  top_pods: 'Top Pods',
+  resource_capacity: 'Resource Capacity',
+  resource_usage: 'Resource Allocation',
+  compute_overview: 'Compute Overview',
+  node_status: 'Node Status',
+
+  // Events
+  event_stream: 'Event Stream',
+  event_summary: 'Event Summary',
+  warning_events: 'Warning Events',
+  recent_events: 'Recent Events',
+  events_timeline: 'Events Timeline',
+
+  // Trend cards
+  pod_health_trend: 'Pod Health Trend',
+  resource_trend: 'Resource Trend',
+
+  // Storage and network
+  storage_overview: 'Storage Overview',
+  pvc_status: 'PVC Status',
+  pv_status: 'PV Status',
+  resource_quota_status: 'Resource Quota Status',
+  network_overview: 'Network Overview',
+  service_status: 'Service Status',
+  ingress_status: 'Ingress Status',
+  network_policy_status: 'Network Policy Status',
+
+  // Namespace cards
+  namespace_overview: 'Namespace Overview',
+  namespace_analysis: 'Namespace Analysis',
+  namespace_rbac: 'Namespace RBAC',
+  namespace_quotas: 'Namespace Quotas',
+  namespace_events: 'Namespace Events',
+  namespace_monitor: 'Namespace Monitor',
+
+  // Operator cards
+  operator_status: 'Operator Status',
+  operator_subscriptions: 'Operator Subscriptions',
+  operator_subscription_status: 'Operator Subscription Status',
+  crd_health: 'CRD Health',
+  configmap_status: 'ConfigMap Status',
+
+  // Helm/GitOps cards
+  gitops_drift: 'GitOps Drift',
+  helm_release_status: 'Helm Release Status',
+  helm_releases: 'Helm Releases',
+  helm_history: 'Helm History',
+  helm_values_diff: 'Helm Values Diff',
+  kustomization_status: 'Kustomization Status',
+  buildpacks_status: 'Buildpacks Status',
+  overlay_comparison: 'Overlay Comparison',
+  chart_versions: 'Helm Chart Versions',
+
+  // ArgoCD cards
+  argocd_applications: 'ArgoCD Applications',
+  argocd_sync_status: 'ArgoCD Sync Status',
+  argocd_health: 'ArgoCD Health',
+
+  // GPU and hardware cards
+  gpu_overview: 'GPU Overview',
+  gpu_status: 'GPU Status',
+  gpu_inventory: 'GPU Inventory',
+  gpu_workloads: 'GPU Workloads',
+  gpu_utilization: 'GPU Utilization',
+  gpu_usage_trend: 'GPU Usage Trend',
+  gpu_namespace_allocations: 'GPU Namespace Allocations',
+  hardware_health: 'Hardware Health',
+
+  // Security, RBAC, and compliance
+  security_issues: 'Security Issues',
+  rbac_overview: 'RBAC Overview',
+  policy_violations: 'Policy Violations',
+  opa_policies: 'OPA Policies',
+  kyverno_policies: 'Kyverno Policies',
+  falco_alerts: 'Falco Alerts',
+  trivy_scan: 'Trivy Scan',
+  kubescape_scan: 'Kubescape Scan',
+  compliance_score: 'Compliance Score',
+  vault_secrets: 'Vault Secrets',
+  external_secrets: 'External Secrets',
+  cert_manager: 'Cert Manager',
+
+  // Alerting cards
+  active_alerts: 'Active Alerts',
+  alert_rules: 'Alert Rules',
+
+  // Cost management
+  opencost_overview: 'OpenCost Overview',
+  kubecost_overview: 'Kubecost Overview',
+
+  // MCS (Multi-Cluster Service) cards
+  service_exports: 'Service Exports',
+  service_imports: 'Service Imports',
+  gateway_status: 'Gateway Status',
+  service_topology: 'Service Topology',
+
+  // Other
+  upgrade_status: 'Cluster Upgrade Status',
+  user_management: 'User Management',
+  github_activity: 'GitHub Activity',
+  kubectl: 'Kubectl Terminal',
+  weather: 'Weather',
+  rss_feed: 'RSS Feed',
+  iframe_embed: 'Iframe Embed',
+  network_utils: 'Network Utils',
+  mobile_browser: 'Mobile Browser',
+
+  // AI cards
+  console_ai_issues: 'AI Issues',
+  console_ai_kubeconfig_audit: 'AI Kubeconfig Audit',
+  console_ai_health_check: 'AI Health Check',
+  console_ai_offline_detection: 'Predictive Health Monitor',
+
+  // Stock Market Ticker
+  stock_market_ticker: 'Stock Market Ticker',
+
+  // PROW CI/CD cards
+  prow_jobs: 'PROW Jobs',
+  prow_status: 'PROW Status',
+  prow_history: 'PROW History',
+
+  // ML/AI workload cards
+  llm_inference: 'llm-d Inference',
+  llm_models: 'llm-d Models',
+  llmd_flow: 'llm-d Request Flow',
+  llmd_ai_insights: 'llm-d AI Insights',
+  llmd_configurator: 'llm-d Configurator',
+  kvcache_monitor: 'KV Cache Monitor',
+  epp_routing: 'EPP Routing',
+  pd_disaggregation: 'P/D Disaggregation',
+  ml_jobs: 'ML Jobs',
+  ml_notebooks: 'ML Notebooks',
+
+  // Runtime cards
+  wasmcloud_status: 'WasmCloud Status',
+
+  // Benchmark cards
+  nightly_e2e_status: 'Nightly E2E Status',
+  benchmark_hero: 'Latest Benchmark',
+  pareto_frontier: 'Performance Explorer',
+  hardware_leaderboard: 'Hardware Leaderboard',
+  latency_breakdown: 'Latency Breakdown',
+  throughput_comparison: 'Throughput Comparison',
+  performance_timeline: 'Performance Timeline',
+  resource_utilization: 'Resource Utilization',
+
+  // Games
+  sudoku_game: 'Sudoku Game',
+  match_game: 'Kube Match',
+  solitaire: 'Kube Solitaire',
+  checkers: 'AI Checkers',
+  game_2048: 'Kube 2048',
+  kubedle: 'Kubedle',
+  pod_sweeper: 'Pod Sweeper',
+  container_tetris: 'Container Tetris',
+  flappy_pod: 'Flappy Pod',
+  kube_man: 'Kube-Man',
+  kube_kong: 'Kube Kong',
+  pod_pitfall: 'Pod Pitfall',
+  node_invaders: 'Node Invaders',
+  pod_crosser: 'Pod Crosser',
+  pod_brothers: 'Pod Brothers',
+  kube_kart: 'Kube Kart',
+  kube_pong: 'Kube Pong',
+  kube_snake: 'Kube Snake',
+  kube_galaga: 'Kube Galaga',
+  kube_doom: 'Kube Doom',
+  kube_craft: 'Kube Craft',
+  kube_chess: 'Kube Chess',
+
+  // Provider health
+  provider_health: 'Provider Health',
+  // CoreDNS
+  coredns_status: 'CoreDNS',
+}
+
+// Short descriptions shown via info icon tooltip in the card header
+export const CARD_DESCRIPTIONS: Record<string, string> = {
+  cluster_health: 'Overall health status of all connected Kubernetes clusters.',
+  cluster_focus: 'Deep-dive view of a single cluster with key metrics and resources.',
+  cluster_network: 'Network connectivity and traffic flow between clusters.',
+  cluster_comparison: 'Side-by-side comparison of clusters by resource usage and health.',
+  cluster_costs: 'Estimated infrastructure costs broken down by cluster.',
+  cluster_metrics: 'Real-time CPU, memory, and pod metrics across clusters.',
+  cluster_locations: 'Geographic map of cluster locations worldwide.',
+  cluster_resource_tree: 'Hierarchical tree view of all resources in a cluster.',
+  app_status: 'Status of workloads across clusters with health indicators.',
+  workload_deployment: 'Deploy workloads to clusters using drag-and-drop.',
+  deployment_missions: 'Track multi-cluster deployment missions and their progress.',
+  deployment_progress: 'Real-time deployment rollout progress and status.',
+  deployment_status: 'Detailed status of deployments including replicas and conditions.',
+  deployment_issues: 'Active deployment problems such as failed rollouts or image pull errors.',
+  statefulset_status: 'Status of StatefulSets including replicas and volume claims.',
+  daemonset_status: 'Status of DaemonSets including node coverage and update progress.',
+  replicaset_status: 'Status of ReplicaSets including replica counts and conditions.',
+  job_status: 'Status of Jobs including completion, duration, and failures.',
+  cronjob_status: 'Status of CronJobs including schedules and recent runs.',
+  hpa_status: 'Status of Horizontal Pod Autoscalers and scaling metrics.',
+  cluster_groups: 'Organize clusters into logical groups for targeted deployments.',
+  resource_marshall: 'Explore resource dependency trees and ownership chains.',
+  workload_monitor: 'Monitor all resources for a workload with health status, alerts, and AI diagnose/repair.',
+  llmd_stack_monitor: 'Monitor the llm-d inference stack: model serving, EPP, gateways, and autoscalers.',
+  prow_ci_monitor: 'Monitor PROW CI jobs with success rates, failure analysis, and AI repair.',
+  github_ci_monitor: 'Monitor GitHub Actions workflows across repos with pass rates and alerts.',
+  cluster_health_monitor: 'Monitor cluster health across all connected clusters with pod and deployment issues.',
+  pod_issues: 'Pods with errors, restarts, or scheduling problems.',
+  top_pods: 'Top resource-consuming pods ranked by CPU or memory usage.',
+  resource_capacity: 'Cluster resource capacity vs. current allocation.',
+  resource_usage: 'CPU and memory allocation breakdown across clusters.',
+  node_status: 'Status of Kubernetes nodes including conditions and capacity.',
+  compute_overview: 'Summary of compute resources: nodes, CPUs, and memory.',
+  event_stream: 'Live stream of Kubernetes events from all clusters.',
+  event_summary: 'Aggregated event counts grouped by type and reason.',
+  warning_events: 'Warning-level events that may need attention.',
+  recent_events: 'Most recent events across all clusters.',
+  events_timeline: 'Timeline chart of event frequency over time.',
+  pod_health_trend: 'Historical trend of pod health status over time.',
+  resource_trend: 'Resource usage trends showing CPU and memory over time.',
+  storage_overview: 'Persistent volume and storage class overview.',
+  pvc_status: 'Status of Persistent Volume Claims across clusters.',
+  pv_status: 'Status of Persistent Volumes including capacity and binding.',
+  resource_quota_status: 'Resource quota utilization and limits across namespaces.',
+  network_overview: 'Network policies, services, and ingress summary.',
+  service_status: 'Status of Kubernetes services and their endpoints.',
+  ingress_status: 'Status of Ingress resources including hosts and backends.',
+  network_policy_status: 'Status of Network Policies and affected pods.',
+  namespace_overview: 'Summary of resources within a namespace.',
+  namespace_analysis: 'Detailed analysis of namespace health and resource usage.',
+  namespace_rbac: 'RBAC roles and bindings within a namespace.',
+  namespace_quotas: 'Resource quota utilization within a namespace.',
+  namespace_events: 'Events filtered to a specific namespace.',
+  namespace_monitor: 'Real-time monitoring of namespace resource trends.',
+  operator_status: 'Status of installed Kubernetes operators.',
+  operator_subscriptions: 'Operator subscriptions and update channels.',
+  operator_subscription_status: 'Detailed status of Operator Lifecycle Manager subscriptions.',
+  crd_health: 'Health and status of Custom Resource Definitions.',
+  configmap_status: 'Status of ConfigMaps including size and update times.',
+  gitops_drift: 'Drift detection between Git source and live cluster state.',
+  helm_release_status: 'Status of Helm releases across clusters.',
+  helm_releases: 'List of all deployed Helm releases.',
+  helm_history: 'Revision history and rollback options for Helm releases.',
+  helm_values_diff: 'Diff of Helm values between revisions.',
+  kustomization_status: 'Status of Kustomize overlays and their resources.',
+  overlay_comparison: 'Compare Kustomize overlays across environments.',
+  chart_versions: 'Available Helm chart versions and update status.',
+  argocd_applications: 'ArgoCD application inventory and sync status.',
+  argocd_sync_status: 'Sync status of ArgoCD-managed applications.',
+  argocd_health: 'Health of ArgoCD applications and components.',
+  gpu_overview: 'Summary of GPU resources across all clusters.',
+  gpu_status: 'Current GPU utilization and health status.',
+  gpu_inventory: 'Inventory of GPU nodes with model, memory, and driver info.',
+  gpu_workloads: 'Workloads running on GPU-enabled nodes.',
+  gpu_utilization: 'Real-time GPU utilization percentage and temperature.',
+  gpu_usage_trend: 'Historical GPU usage trends over time.',
+  hardware_health: 'Detects hardware device disappearances (GPUs, NICs, NVMe, InfiniBand) that often require a power cycle to recover. Common with SuperMicro/HGX systems. Also shows full device inventory per node.',
+  security_issues: 'Security vulnerabilities and misconfigurations detected.',
+  rbac_overview: 'Overview of RBAC roles, bindings, and permissions.',
+  policy_violations: 'Active policy violations from OPA, Kyverno, or other engines.',
+  opa_policies: 'OPA Gatekeeper policies and constraint status.',
+  kyverno_policies: 'Kyverno policies and their enforcement status.',
+  falco_alerts: 'Runtime security alerts from Falco.',
+  trivy_scan: 'Container image vulnerability scan results from Trivy.',
+  kubescape_scan: 'Security posture scan results from Kubescape.',
+  compliance_score: 'Overall compliance score across security frameworks.',
+  vault_secrets: 'HashiCorp Vault secrets management status.',
+  external_secrets: 'External Secrets Operator sync status.',
+  cert_manager: 'TLS certificate status and renewal from cert-manager.',
+  active_alerts: 'Currently firing alerts from Prometheus or other sources.',
+  alert_rules: 'Configured alert rules and their evaluation status.',
+  opencost_overview: 'Cost allocation data from OpenCost.',
+  kubecost_overview: 'Cost breakdown and optimization from Kubecost.',
+  service_exports: 'Services exported for multi-cluster discovery.',
+  service_imports: 'Services imported from other clusters.',
+  gateway_status: 'Gateway API resource status and routing.',
+  service_topology: 'Visual topology of service-to-service communication.',
+  upgrade_status: 'Kubernetes version upgrade status and available upgrades.',
+  user_management: 'Manage console users and their roles.',
+  github_activity: 'Recent GitHub activity: commits, PRs, and issues.',
+  kubectl: 'Interactive kubectl terminal for running commands.',
+  weather: 'Current weather conditions for cluster locations.',
+  rss_feed: 'RSS feed reader for Kubernetes news and blogs.',
+  iframe_embed: 'Embed an external web page inside a card.',
+  network_utils: 'Network diagnostic utilities: ping, DNS, traceroute.',
+  mobile_browser: 'Embedded mobile-sized browser for testing.',
+  console_ai_issues: 'AI-detected issues and recommended fixes.',
+  console_ai_kubeconfig_audit: 'AI audit of kubeconfig files for security and cleanup.',
+  console_ai_health_check: 'AI-powered cluster health analysis.',
+  console_ai_offline_detection: 'Monitors cluster health and predicts failures before they happen. Detects offline nodes, GPU exhaustion, resource pressure, and groups issues by root cause for efficient remediation.',
+  stock_market_ticker: 'Live stock market ticker with tech company prices.',
+  prow_jobs: 'PROW CI/CD job status and results.',
+  prow_status: 'Overall PROW system health and queue depth.',
+  prow_history: 'Historical PROW job runs and success rates.',
+  llm_inference: 'llm-d inference endpoint status and request metrics.',
+  llm_models: 'LLM models deployed via llm-d with version info.',
+  llmd_flow: 'Animated visualization of inference request flow through the llm-d stack: load balancer → EPP → prefill/decode pods.',
+  llmd_ai_insights: 'AI-generated insights about llm-d performance, bottlenecks, and optimization recommendations.',
+  llmd_configurator: 'Configure llm-d deployment parameters: replicas, autoscaling, model variants, and resource limits.',
+  kvcache_monitor: 'Real-time KV cache utilization across inference pods with hit rates and memory usage.',
+  epp_routing: 'Endpoint Picker Pod routing decisions: how requests are distributed based on KV cache affinity.',
+  pd_disaggregation: 'Prefill/Decode disaggregation architecture: separate pools for prompt processing and token generation.',
+  ml_jobs: 'Machine learning training and batch job status.',
+  ml_notebooks: 'Jupyter notebook server status and resource usage.',
+  provider_health: 'Health and status of AI and cloud infrastructure providers.',
+
+  // Runtime cards
+  wasmcloud_status: 'wasmCloud host status, actor inventory, and health monitoring.',
+
+  // Games
+  sudoku_game: 'Classic Sudoku puzzle game with multiple difficulty levels.',
+  match_game: 'Memory matching game with Kubernetes resource icons.',
+  solitaire: 'Classic Klondike solitaire card game.',
+  checkers: 'Play checkers against an AI opponent.',
+  game_2048: 'Slide and merge tiles to reach 2048.',
+  kubedle: 'Wordle-style game with Kubernetes terminology.',
+  pod_sweeper: 'Minesweeper clone with a Kubernetes pod theme.',
+  container_tetris: 'Classic Tetris with container-shaped blocks.',
+  flappy_pod: 'Navigate a pod through cluster obstacles.',
+  kube_man: 'Pac-Man style game collecting resources in a cluster maze.',
+  kube_kong: 'Donkey Kong inspired platformer with Kubernetes theme.',
+  pod_pitfall: 'Pitfall-style adventure game as a pod.',
+  node_invaders: 'Space Invaders clone defending your cluster.',
+  pod_crosser: 'Frogger-style game crossing cluster traffic.',
+  pod_brothers: 'Super Mario Bros inspired platformer.',
+  kube_kart: 'Racing game through Kubernetes infrastructure.',
+  kube_pong: 'Classic Pong game with cluster theming.',
+  kube_snake: 'Snake game collecting Kubernetes resources.',
+  kube_galaga: 'Galaga-style shooter defending against threats.',
+  kube_doom: 'First-person debugging adventure.',
+  kube_craft: 'Build and manage your cluster world.',
+  kube_chess: 'Chess game with Kubernetes-themed pieces.',
+  // CoreDNS
+  coredns_status: 'CoreDNS pod health, restart counts, and cluster status across clusters.',
+}
